requirements.txt
PyQt5==5.15.10
selenium==4.15.2
openai==1.3.0
requests==2.31.0
python-dotenv==1.0.0
pytz==2023.3